refactor: move Chat class to its own file
